Move test_relations_with_limited_querysets file within test_relations_pk
